Temporarily in this office redundancy will be temporarily permitted.
It seems that somewhere in the build, -march=armv5te is being specified somewhere. This ends up overriding and takes preference over -mcpu=cortex-a9 (which should be equivalent to -march=armv7-a -mtune=cortex-a9 in the first place). Whilst testing and experimenting and testing and experimenting, let's just be temporarily redundant here in this location.